creates a tracker to carousel scroll only when in screen view
