+++ package.json
@@ -8,6 +8,7 @@
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-icons": "^4.2.0",
+    "react-on-screen": "^2.1.1",
     "styled-components": "^5.2.3",
     "typescript": "^4.2.4",
     "web-vitals": "^1.1.1"
